+++ ehtim/imaging/pol_imager_utils.py
@@ -20,15 +20,13 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from ehtim.backends import array_namespace
-
-try:
-    from pynfft.nfft import NFFT
-    _HAS_NFFT = True
-except ImportError:
-    NFFT = None
-    _HAS_NFFT = False
-
-from ehtim.const_def import FFT_PAD_DEFAULT, GRIDDER_P_RAD_DEFAULT, NFFT_EPS_DEFAULT, RADPERAS
+from ehtim.const_def import (
+    FFT_PAD_DEFAULT,
+    GRIDDER_P_RAD_DEFAULT,
+    NFFT_EPS_DEFAULT,
+    NFFT_NTHREADS_DEFAULT,
+    RADPERAS,
+)
 from ehtim.observing.obs_helpers import NFFTInfo, ftmatrix, nufft2_backend, ticks
 
 TANWIDTH_M = 0.5
@@ -1485,6 +1483,7 @@ def chisqdata_pvis_nfft(Obsdata, Prior, **kwargs):
     fft_pad_factor = kwargs.get('fft_pad_factor',FFT_PAD_DEFAULT)
     p_rad = kwargs.get('p_rad', GRIDDER_P_RAD_DEFAULT)
     nfft_eps = kwargs.get('nfft_eps', NFFT_EPS_DEFAULT)
+    nfft_nthreads = kwargs.get('nfft_nthreads', NFFT_NTHREADS_DEFAULT)
 
     # unpack data
     data_arr = Obsdata.unpack(['u','v','pvis','psigma'], conj=True)
@@ -1494,7 +1493,8 @@ def chisqdata_pvis_nfft(Obsdata, Prior, **kwargs):
 
     # get NFFT info
     npad = int(fft_pad_factor * np.max((Prior.xdim, Prior.ydim)))
-    A1 = NFFTInfo(Prior.xdim, Prior.ydim, Prior.psize, Prior.pulse, npad, p_rad, uv, eps=nfft_eps)
+    A1 = NFFTInfo(Prior.xdim, Prior.ydim, Prior.psize, Prior.pulse, npad, p_rad, uv,
+                  eps=nfft_eps, nthreads=nfft_nthreads)
     A = [A1]
 
     return (vis, sigma, A)
@@ -1522,6 +1522,7 @@ def chisqdata_m_nfft(Obsdata, Prior, **kwargs):
     fft_pad_factor = kwargs.get('fft_pad_factor',FFT_PAD_DEFAULT)
     p_rad = kwargs.get('p_rad', GRIDDER_P_RAD_DEFAULT)
     nfft_eps = kwargs.get('nfft_eps', NFFT_EPS_DEFAULT)
+    nfft_nthreads = kwargs.get('nfft_nthreads', NFFT_NTHREADS_DEFAULT)
 
     # unpack data
     mdata = Obsdata.unpack(['u','v','m','msigma'], conj=True)
@@ -1531,7 +1532,8 @@ def chisqdata_m_nfft(Obsdata, Prior, **kwargs):
 
     # get NFFT info
     npad = int(fft_pad_factor * np.max((Prior.xdim, Prior.ydim)))
-    A1 = NFFTInfo(Prior.xdim, Prior.ydim, Prior.psize, Prior.pulse, npad, p_rad, uv, eps=nfft_eps)
+    A1 = NFFTInfo(Prior.xdim, Prior.ydim, Prior.psize, Prior.pulse, npad, p_rad, uv,
+                  eps=nfft_eps, nthreads=nfft_nthreads)
     A = [A1]
 
     return (m, sigmam, A)
@@ -1559,6 +1561,7 @@ def chisqdata_vvis_nfft(Obsdata, Prior, **kwargs):
     fft_pad_factor = kwargs.get('fft_pad_factor',FFT_PAD_DEFAULT)
     p_rad = kwargs.get('p_rad', GRIDDER_P_RAD_DEFAULT)
     nfft_eps = kwargs.get('nfft_eps', NFFT_EPS_DEFAULT)
+    nfft_nthreads = kwargs.get('nfft_nthreads', NFFT_NTHREADS_DEFAULT)
 
     # unpack data
     data_arr = Obsdata.unpack(['u','v','vvis','vsigma'], conj=False)
@@ -1568,7 +1571,8 @@ def chisqdata_vvis_nfft(Obsdata, Prior, **kwargs):
 
     # get NFFT info
     npad = int(fft_pad_factor * np.max((Prior.xdim, Prior.ydim)))
-    A1 = NFFTInfo(Prior.xdim, Prior.ydim, Prior.psize, Prior.pulse, npad, p_rad, uv, eps=nfft_eps)
+    A1 = NFFTInfo(Prior.xdim, Prior.ydim, Prior.psize, Prior.pulse, npad, p_rad, uv,
+                  eps=nfft_eps, nthreads=nfft_nthreads)
     A = [A1]
 
     return (vis, sigma, A)